backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pdf2image==1.17.0
Pillow==10.4.0
python-multipart==0.0.9
cloudinary==1.41.0
python-dotenv==1.0.1
